+++ engine/panicx/doc.go
@@ -1,0 +1,3 @@
+// Package panicx contains utilities for providing meaningful contexts to panics
+// that occur with the engine.
+package panicx